✨ feat: Add useSearchParams hook to SignIn component
This commit adds the useSearchParams hook to the SignIn component in order to retrieve the email parameter from the URL query string. The retrieved email is then used as the default value for the email input field in the sign-in form. This change improves the user experience by pre-filling the email field for users who are redirected to the sign-in page with an email parameter in the URL.