🚀 feat: Update npm dependencies to latest versions
This commit updates the npm dependencies in the package.json file to their latest stable versions. The updated dependencies include:
- "@hookform/resolvers": "^3.3.2"
- "axios": "^1.6.5"
- "clsx": "^2.0.0"
- "date-fns": "^3.2.0"
- "lucide-react": "^0.295.0"
- "match-sorter": "^6.3.1"
- "react-day-picker": "^8.10.0"
- "react-helmet-async": "^2.0.3"
- "react-hook-form": "^7.49.2"
- "react-router-dom": "^6.21.0"
- "recharts": "^2.10.3"
- "sonner": "^1.2.4"
- "tailwind-merge": "^2.1.0"
- "zod": "^3.22.4"

Updating the dependencies ensures that the project is using the latest features, bug fixes, and security patches provided by the library maintainers. This helps to keep the codebase up-to-date and maintain compatibility with other dependencies.